requirements.txt
Numpy==1.6.1
Google-cloud==0.32.0
google-api-python-client
Flask==0.12.2
google-cloud-datastore==1.4.0